requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
python-multipart==0.0.9
ezdxf==1.3.3
numpy==1.26.4
numpy-stl==3.1.1
shapely==2.0.3
mapbox-earcut==1.0.1
aiofiles==23.2.1
python-dotenv>=1.0.0
google-genai>=1.0.0
